+++ src/sized_chunk.rs
@@ -76,7 +76,6 @@ use crate::types::ChunkLength;
 /// # extern crate typenum;
 /// # use sized_chunks::Chunk;
 /// # use typenum::U64;
-/// # fn main() {
 /// // Construct a chunk with a 64 item capacity
 /// let mut chunk = Chunk::<i32, U64>::new();
 /// // Fill it with descending numbers
@@ -86,7 +85,6 @@ use crate::types::ChunkLength;
 /// // It's got all the amenities like `FromIterator` and `Eq`
 /// let expected: Chunk<i32, U64> = (0..64).collect();
 /// assert_eq!(expected, chunk);
-/// # }
 /// ```
 ///
 /// [Unsigned]: https://docs.rs/typenum/1.10.0/typenum/marker_traits/trait.Unsigned.html
@@ -971,6 +969,72 @@ where
     A: 'a,
     N: ChunkLength<A> + 'a,
 {
+}
+
+#[cfg(feature = "refpool")]
+mod refpool {
+    use super::*;
+    use ::refpool::{PoolClone, PoolDefault};
+    use std::mem::MaybeUninit;
+
+    impl<A, N> PoolDefault for Chunk<A, N>
+    where
+        N: ChunkLength<A>,
+    {
+        unsafe fn default_uninit(target: &mut MaybeUninit<Self>) {
+            let ptr = target.as_mut_ptr();
+            let left_ptr: *mut usize = &mut (*ptr).left;
+            let right_ptr: *mut usize = &mut (*ptr).right;
+            left_ptr.write(0);
+            right_ptr.write(0);
+        }
+    }
+
+    impl<A, N> PoolClone for Chunk<A, N>
+    where
+        A: Clone,
+        N: ChunkLength<A>,
+    {
+        unsafe fn clone_uninit(&self, target: &mut MaybeUninit<Self>) {
+            let ptr = target.as_mut_ptr();
+            let left_ptr: *mut usize = &mut (*ptr).left;
+            let right_ptr: *mut usize = &mut (*ptr).right;
+            let data_ptr: *mut _ = &mut (*ptr).data;
+            let data_ptr: *mut A = (*data_ptr).as_mut_ptr().cast();
+            left_ptr.write(self.left);
+            right_ptr.write(self.right);
+            for index in self.left..self.right {
+                data_ptr.add(index).write((*self.ptr(index)).clone());
+            }
+        }
+    }
+
+    #[cfg(test)]
+    mod test {
+        use super::*;
+        use ::refpool::{Pool, PoolRef};
+
+        #[test]
+        fn default_and_clone() {
+            let mut pool: Pool<Chunk<usize>> = Pool::new(16);
+            let mut ref1 = PoolRef::default(&mut pool);
+            {
+                let chunk = PoolRef::make_mut(&mut pool, &mut ref1);
+                chunk.push_back(1);
+                chunk.push_back(2);
+                chunk.push_back(3);
+            }
+            let ref2 = ref1.cloned(&mut pool);
+            let ref3 = PoolRef::clone_from(&mut pool, &Chunk::from_iter(1..=3));
+            assert_eq!(Chunk::<usize>::from_iter(1..=3), *ref1);
+            assert_eq!(Chunk::<usize>::from_iter(1..=3), *ref2);
+            assert_eq!(Chunk::<usize>::from_iter(1..=3), *ref3);
+            assert_eq!(ref1, ref2);
+            assert_eq!(ref1, ref3);
+            assert_eq!(ref2, ref3);
+            assert!(!PoolRef::ptr_eq(&ref1, &ref2));
+        }
+    }
 }
 
 #[cfg(test)]