+++ src/sparse_chunk.rs
@@ -41,7 +41,6 @@ use crate::types::ChunkLength;
 /// # extern crate typenum;
 /// # use sized_chunks::SparseChunk;
 /// # use typenum::U20;
-/// # fn main() {
 /// // Construct a chunk with a 20 item capacity
 /// let mut chunk = SparseChunk::<i32, U20>::new();
 /// // Set the 18th index to the value 5.
@@ -53,7 +52,6 @@ use crate::types::ChunkLength;
 /// assert_eq!(chunk.get(5), Some(&23));
 /// assert_eq!(chunk.get(6), None);
 /// assert_eq!(chunk.get(18), Some(&5));
-/// # }
 /// ```
 ///
 /// [Unsigned]: https://docs.rs/typenum/1.10.0/typenum/marker_traits/trait.Unsigned.html
@@ -258,6 +256,12 @@ where
     }
 }
 
+impl<A, N: Bits + ChunkLength<A>> Default for SparseChunk<A, N> {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
 impl<A, N: Bits + ChunkLength<A>> Index<usize> for SparseChunk<A, N> {
     type Output = A;
 
@@ -405,6 +409,62 @@ impl<'a, A, N: Bits + ChunkLength<A>> Iterator for Drain<A, N> {
 
     fn next(&mut self) -> Option<Self::Item> {
         self.chunk.pop()
+    }
+}
+
+#[cfg(feature = "refpool")]
+mod refpool {
+    use super::*;
+    use ::refpool::{PoolClone, PoolDefault};
+    use std::mem::MaybeUninit;
+
+    impl<A, N> PoolDefault for SparseChunk<A, N>
+    where
+        N: Bits + ChunkLength<A>,
+    {
+        unsafe fn default_uninit(target: &mut MaybeUninit<Self>) {
+            let ptr = target.as_mut_ptr();
+            let map_ptr: *mut Bitmap<N> = &mut (*ptr).map;
+            map_ptr.write(Bitmap::new());
+        }
+    }
+
+    impl<A, N> PoolClone for SparseChunk<A, N>
+    where
+        A: Clone,
+        N: Bits + ChunkLength<A>,
+    {
+        unsafe fn clone_uninit(&self, target: &mut MaybeUninit<Self>) {
+            let ptr = target.as_mut_ptr();
+            let map_ptr: *mut Bitmap<N> = &mut (*ptr).map;
+            let data_ptr: *mut _ = &mut (*ptr).data;
+            let data_ptr: *mut A = (*data_ptr).as_mut_ptr().cast();
+            map_ptr.write(self.map);
+            for index in &self.map {
+                data_ptr.add(index).write(self[index].clone());
+            }
+        }
+    }
+
+    #[cfg(test)]
+    mod test {
+        use super::*;
+        use ::refpool::{Pool, PoolRef};
+
+        #[test]
+        fn default_and_clone() {
+            let mut pool: Pool<SparseChunk<usize>> = Pool::new(16);
+            let mut ref1 = PoolRef::default(&mut pool);
+            {
+                let chunk = PoolRef::make_mut(&mut pool, &mut ref1);
+                chunk.insert(5, 13);
+                chunk.insert(10, 37);
+                chunk.insert(31, 337);
+            }
+            let ref2 = ref1.cloned(&mut pool);
+            assert_eq!(ref1, ref2);
+            assert!(!PoolRef::ptr_eq(&ref1, &ref2));
+        }
     }
 }